+++ genvsproj.cmd
@@ -1,5 +1,6 @@
-set STEAMDIR="C:\\Program Files (x86)\\Steam\\steamapps\\common\\The Scroll Of Taiwu\\"
+::set STEAMDIR="C:\\Program Files (x86)\\Steam\\steamapps\\common\\The Scroll Of Taiwu\\"
 ::set STEAMDIR="E:\\SteamLibrarys\\steamapps\\common\\The Scroll Of Taiwu\\"
+set STEAMDIR="D:\Software\steam\steamapps\common\The Scroll Of Taiwu"
 mkdir build
 cd build
